+++ breeze-core/src/test/kotlin/com/windea/breezeframework/core/extensions/AnnotationExtensionsKtTest.kt
@@ -1,3 +1,5 @@
 package com.windea.breezeframework.core.extensions
 
 class AnnotationExtensionsKtTest
+
+object Foo